Create an interface which the configuration properties implements
